+++ src/org/isf/examination/test/TestPatientExamination.java
@@ -12,7 +12,7 @@ import org.isf.utils.exception.OHException;
 public class TestPatientExamination 
 {	
 	private Timestamp pex_date = new Timestamp(1000);
-	private int pex_height = 0;	
+	private double pex_height = 0;	
 	private int pex_weight = 60;	
 	private int pex_pa_min = 80;	
 	private int pex_pa_max = 120;
